Refactored NUMERIC aggregate routing helpers.
Moved has_numeric_aggregate_or_group onto CreateUpperPathsHookArgs as
has_numeric_aggregate, and inlined numeric_agg_expr at its call sites
using AggregateUDF::call.
